+++ packages/sampler/stories/default/Picker.js
@@ -2,9 +2,9 @@ import Picker from '@enact/moonstone/Picker';
 import {decrementIcons, incrementIcons} from './icons';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
-import {action} from '@storybook/addon-actions';
 
 import {boolean, select, text} from '../../src/enact-knobs';
+import {action} from '../../src/utils';
 
 // Set up some defaults for info and knobs
 const prop = {