+++ packages/sampler/stories/qa/ExpandableInput.js
@@ -2,9 +2,9 @@ import ExpandableInput from '@enact/moonstone/ExpandableInput';
 import {icons} from '@enact/moonstone/Icon';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
-import {action} from '@storybook/addon-actions';
 
 import {boolean, select, text} from '../../src/enact-knobs';
+import {action} from '../../src/utils';
 
 const iconNames = ['', ...Object.keys(icons)];