+++ packages/spotlight/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `spotlight` to attempt to restore focus to an element nearest the pointer position when the pointer hides within an overflow container
+
 ## [3.0.0-rc.2] - 2019-08-08
 
 No significant changes.